Remove commented code in GraphicalPlayer
Some code was left commented in order to preserve the previous version
of the code we were too afraid to delete, this was unnecessary.

This commit deletes the useless code in order to clean up things.
The next additions should be further work to replace listeners with
observable properties on maps.